+++ app/src/main/java/com/forrestguice/suntimeswidget/SuntimesUtils.java
@@ -20,6 +20,7 @@ package com.forrestguice.suntimeswidget;
 
 import android.content.Context;
 
+import android.content.res.ColorStateList;
 import android.graphics.Bitmap;
 import android.graphics.Canvas;
 import android.graphics.LightingColorFilter;
@@ -1542,6 +1543,36 @@ public class SuntimesUtils
                 }
             }
         }
+    }
+
+    /**
+     * @param enabledColor normal state color
+     * @param disabledColor disabled state color
+     * @return a ColorStateList w/ enabled / disabled states (intended for text label)
+     */
+    public static ColorStateList colorStateList(int enabledColor, int disabledColor)
+    {
+        return new ColorStateList(
+                new int[][] { new int[] { android.R.attr.state_enabled}, new int[] {-android.R.attr.state_enabled}},
+                new int[] {enabledColor, disabledColor}
+        );
+    }
+
+    /**
+     * @param enabledColor normal state color
+     * @param disabledColor disabled state color
+     * @param pressedColor pressed/focused color
+     * @return a ColorStateList w/ pressed, enabled, and disabled states (intended for text button)
+     */
+    public static ColorStateList colorStateList(int enabledColor, int disabledColor, int pressedColor)
+    {
+        return new ColorStateList(
+                new int[][] { new int[] { android.R.attr.state_pressed},
+                        new int[] { android.R.attr.state_focused},
+                        new int[] {-android.R.attr.state_enabled},
+                        new int[] {} },
+                new int[] {pressedColor, enabledColor, disabledColor, enabledColor}
+        );
     }
 
 }
